pyproject.toml
[tool.poetry]
name = "blogapi"
version = "0.1.0"
description = ""
authors = ["SherlockH0 <ddmlf7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"
psycopg2 = "^2.9.10"
django-ninja = "^1.3.0"
django-ninja-jwt = "^5.3.4"
django-ninja-extra = "^0.21.4"
isodate = "^0.7.2"

[tool.poetry.group.dev.dependencies]
django-types = "^0.19.1"
pytest-django = "^4.9.0"
pytest-cov = "^5.0.0"
model-bakery = "^1.20.0"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "blogapi.project.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["blogapi"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"

[flake8]
max-line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
